refactor(api-gateway): split createApp + remove async route-mount race
server.ts createApp (907 LOC) decomposed into:
- app/persistence.ts — DB selection + repo wiring
- app/auth-routes.ts — login/OAuth/SAML/setup/current-user
- app/rbac-routes.ts — admin/orgs/teams/SAs/access-control/folders
- app/domain-routes.ts — chat/dashboards/alerts/datasources/investigations
- app/lifecycle.ts — shutdown hooks

createApp is now end-to-end async; the IIFE that previously raced
route registration against incoming requests is gone. The matching
503 'auth subsystem not ready' shim in middleware/auth.ts is
removed (a boot bug should fail loudly, not 503 user requests).

server.ts down to ~190 LOC.

Inline RBAC SQL extracted to repo methods: IDashboardRepository
.getFolderUid + IAlertRuleRepository.getFolderUid (sqlite +
event-wrapper + legacy in-memory store implementations).
