+++ packages/api-gateway/src/middleware/auth.ts
@@ -79,12 +79,16 @@ function hashApiKey(plaintext: string): string {
 // — Module-level middleware singleton ——————————————————————————
 //
 // Several existing route files import a default `authMiddleware` function.
-// Rather than rewriting every one of them (outside T2 scope), we expose a
-// lazy wrapper that delegates to the middleware built by `createAuthMiddleware`.
-// `server.ts` registers the real implementation at boot via `setAuthMiddleware`.
+// Rather than rewriting every one of them, we expose a wrapper that
+// delegates to the middleware built by `createAuthMiddleware`.
+// `createApp()` (via `app/auth-routes.ts::buildAuthSubsystem`) registers
+// the real implementation at boot via `setAuthMiddleware` BEFORE any
+// route files are mounted, so the wrapper has always been bound by the
+// time a request arrives.
 //
-// When no implementation has been registered yet, the wrapper 503s — never
-// fall through to a permissive default.
+// If something tries to invoke the middleware before binding, that is a
+// programming error in the boot sequence — fail loudly rather than
+// silently 503'ing a request that should never have happened.
 
 type MW = (
   req: AuthenticatedRequest,
@@ -104,10 +108,9 @@ export function authMiddleware(
   next: NextFunction,
 ): void {
   if (!resolvedMiddleware) {
-    res.status(503).json({
-      error: { code: 'SERVICE_UNAVAILABLE', message: 'auth subsystem not initialised' },
-    });
-    return;
+    throw new Error(
+      'authMiddleware invoked before setAuthMiddleware — boot sequence is broken',
+    );
   }
   void resolvedMiddleware(req, res, next);
 }
