+++ packages/data-layer/src/repository/interfaces.ts
@@ -238,6 +238,11 @@ export interface IDashboardRepository {
   updatePanels(id: string, panels: PanelConfig[]): MaybeAsync<Dashboard | undefined>;
   updateVariables(id: string, variables: DashboardVariable[]): MaybeAsync<Dashboard | undefined>;
   delete(id: string): MaybeAsync<boolean>;
+  /**
+   * Resolve the folder UID for a dashboard within an org. Used by RBAC
+   * resolvers to enforce folder-scoped permissions.
+   */
+  getFolderUid(orgId: string, dashboardId: string): MaybeAsync<string | null>;
 }
 
 // — Conversation
@@ -297,6 +302,11 @@ export interface IAlertRuleRepository {
   findPolicyById(id: string): MaybeAsync<NotificationPolicy | undefined>;
   updatePolicy(id: string, patch: Partial<Omit<NotificationPolicy, 'id' | 'createdAt'>>): MaybeAsync<NotificationPolicy | undefined>;
   deletePolicy(id: string): MaybeAsync<boolean>;
+  /**
+   * Resolve the folder UID for an alert rule within an org. Used by RBAC
+   * resolvers to enforce folder-scoped permissions.
+   */
+  getFolderUid(orgId: string, ruleId: string): MaybeAsync<string | null>;
 }
 
 // — Notification (contact points, policy tree, mute timings)
